remove django pagination IOT make api call correctly
